refactor: rename index.tsx
