+++ .prettierignore
@@ -1,4 +1,7 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+.github
+.vscode
+.husky
 
 # dependencies
 /node_modules
